+++ changelog.md
@@ -1,3 +1,18 @@
+## 2025-10-23
+
+Added styling related and components
+
+### Added
+
+-   Setup theme
+-   Added fontWeights constants
+-   Setup global font using Next.js font optimization
+-   Added NewTaskButton, TaskCard, TaskSection components
+
+### Updated
+
+-   Updated README.md
+
 ## 2025-10-14
 
 Setup project